frontend-app-api: move some createApp tests to createAppGraph
